Refactors quantization for improved performance
Simplifies and improves quantization logic by consolidating
quantization and dequantization functions, enhancing code
maintainability.

Changes embedding vectors to use float64 for improved precision.

Updates the vector and matrix creation to accept raw float64
arrays in addition to quantized byte arrays.

+++ ai/embed.go
@@ -20,7 +20,7 @@ type EmbedResponse struct {
 // Embedding represents a single embedding
 type Embedding struct {
 	Index     int       `json:"index"`
-	Embedding []float32 `json:"embedding"`
+	Embedding []float64 `json:"embedding"`
 }
 
 // Embed generates embeddings
